review updates pt.1

+++ scripts/android/build_mwebd.sh
@@ -1,10 +1,13 @@
-# install go > 1.23:
-wget https://go.dev/dl/go1.23.1.linux-amd64.tar.gz
-sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf go1.23.1.linux-amd64.tar.gz
-export PATH=$PATH:/usr/local/go/bin
-export PATH=$PATH:~/go/bin
-go install golang.org/x/mobile/cmd/gomobile@latest
-gomobile init
+if [[ "$1" == "--install" ]]; then
+  # install go > 1.23:
+  wget https://go.dev/dl/go1.23.1.linux-amd64.tar.gz
+  sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf go1.23.1.linux-amd64.tar.gz
+  export PATH=$PATH:/usr/local/go/bin
+  export PATH=$PATH:~/go/bin
+  go install golang.org/x/mobile/cmd/gomobile@latest
+  gomobile init
+fi
+
 # build mwebd:
 git clone https://github.com/ltcmweb/mwebd
 cd mwebd
